Rename views-design to views-material
This breaks the public API since the package is also renamed, but it's
not a big deal compared to AndroidX migration, and we're still in alpha.
