+++ settings.gradle
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-include ':sample', ':typesaferecyclerview', ':selectableviews', ':checkedlazy', ':appctx', ':preferences', ':stetho-init', ':material-lists', ':bundle', ':initprovider', ':exceptions', ':uithread', ':systemservices', ':resources', ':fragmentargs', ':arch-lifecycle', ':mainhandler', ':arch-room', ':bitflags', ':toast', ':snackbar', ':dimensions', ':views', ':views-appcompat', ':viewdsl', ':viewdsl-appcompat', ':viewdsl-constraintlayout', ':viewdsl-design', ':alertdialog', ':alertdialog-appcompat', ':viewdsl-appcompat-styles', ':selectableviews-appcompat', ':selectableviews-constraintlayout', ':viewdsl-ide-preview', ':viewdsl-recyclerview', ':views-cardview', ':material-colors', ':views-recyclerview', ':views-design', ':activities', ':fragments', ':intents', ':collections'
+include ':sample', ':typesaferecyclerview', ':selectableviews', ':checkedlazy', ':appctx', ':preferences', ':stetho-init', ':material-lists', ':bundle', ':initprovider', ':exceptions', ':uithread', ':systemservices', ':resources', ':fragmentargs', ':arch-lifecycle', ':mainhandler', ':arch-room', ':bitflags', ':toast', ':snackbar', ':dimensions', ':views', ':views-appcompat', ':viewdsl', ':viewdsl-appcompat', ':viewdsl-constraintlayout', ':viewdsl-design', ':alertdialog', ':alertdialog-appcompat', ':viewdsl-appcompat-styles', ':selectableviews-appcompat', ':selectableviews-constraintlayout', ':viewdsl-ide-preview', ':viewdsl-recyclerview', ':views-cardview', ':material-colors', ':views-recyclerview', ':views-material', ':activities', ':fragments', ':intents', ':collections'
