+++ views-material/src/main/AndroidManifest.xml
@@ -13,4 +13,4 @@
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
   -->
-<manifest package="splitties.views.design"/>
+<manifest package="splitties.views.material"/>
